rewrite the abstract class
