+++ DiverseSelector/Optisim.py
@@ -1,8 +1,7 @@
+from DiverseSelector.base import SelectionBase
 import numpy as np
-from sklearn.datasets import make_blobs
 
-
-class OptiSim:
+class OptiSim(SelectionBase):
     def __init__(self, r=None, k=10):
         self.r = r
         self.k = k
@@ -78,77 +77,3 @@ class OptiSim:
             return result
         else:
             return self.optisim(arr)
-
-    def select(self, arr, num_selected, func_distance=None, labels=None):
-        """
-         MinMax algorithm for selecting points.
-
-        Parameters
-        ----------
-        arr: np.ndarray
-            Array of features if fun_distance is provided.
-            Otherwise, treated as distance matrix.
-        func_distance: callable
-            function for calculating the pairwise distance between instances of the array.
-            Default is None.
-        num_selected: int
-            number of points that need to be selected
-        labels: np.ndarray
-            labels for performing algorithm withing clusters.
-            Default is None.
-
-        Returns
-        -------
-        selected: list
-            list of ids of selected molecules
-        """
-        self.n_mols = arr.shape[0]
-        if func_distance is not None:
-            self.arr_dist = func_distance(arr)
-        else:
-            self.arr_dist = arr
-
-        if labels is not None:
-            unique_labels = np.unique(labels)
-            num_clusters = len(unique_labels)
-            selected_all = []
-            totally_used = []
-
-            amount_molecules = np.array(
-                [len(np.where(labels == unique_label)[0]) for unique_label in unique_labels])
-
-            n = (num_selected - len(selected_all)) // (num_clusters - len(totally_used))
-
-            while np.any(amount_molecules <= n):
-                for unique_label in unique_labels:
-                    if unique_label not in totally_used:
-                        cluster_ids = np.where(labels == unique_label)[0]
-                        if len(cluster_ids) <= n:
-                            selected_all.append(cluster_ids)
-                            totally_used.append(unique_label)
-
-                n = (num_selected - len(selected_all)) // (num_clusters - len(totally_used))
-                amount_molecules = np.delete(amount_molecules, totally_used)
-
-                warnings.warn(f"Number of molecules in one cluster is less than"
-                              f" {num_selected}/{num_clusters}.\nNumber of selected "
-                              f"molecules might be less than desired.\nIn order to avoid this "
-                              f"problem. Try to use less number of clusters")
-
-            for unique_label in unique_labels:
-                if unique_label not in totally_used:
-                    cluster_ids = np.where(labels == unique_label)[0]
-                    print(cluster_ids.shape)
-                    # arr_dist_cluster = self.arr_dist[cluster_ids]
-                    # selected = self.select_from_cluster(arr_dist_cluster, n)
-                    selected = self.select_from_cluster(self.arr_dist, n, cluster_ids)
-                    selected_all.append(cluster_ids[selected])
-            return np.hstack(selected_all).flatten().tolist()
-        else:
-            selected = self.select_from_cluster(self.arr_dist, num_selected)
-            return selected
-
-# if __name__ == "__main__":
-#     arr, labels = make_blobs(n_samples=100, n_features=2, centers=1, random_state=42)
-#     selected = OptiSim(k=10).select(arr, 12)
-#     print(selected)
