+++ DiverseSelector/base.py
@@ -34,80 +34,101 @@ from DiverseSelector.utils import PandasDataFrame
 import numpy as np
 import pandas as pd
 from sklearn.preprocessing import StandardScaler
+import warnings
 
 
 class SelectionBase(ABC):
     """Base class for subset selection."""
 
-    def __init__(self,
-                 features: Union[np.ndarray, PandasDataFrame, str, PurePath] = None,
-                 arr_dist: np.array = None,
-                 num_selected: int = None,
-                 normalize_features: bool = False,
-                 sep: str = ",",
-                 engine: str = "python",
-                 random_seed: int = 42,
-                 **kwargs,
-                 ):
-        """Abstract class for other modules.
+    def __init__(self):
+        self.arr_dist = None
+        self.n_mols = None
+
+    def select(self, arr, num_selected, func_distance=None, labels=None):
+        """
+         MinMax algorithm for selecting points.
 
         Parameters
         ----------
-        metric : str, optional
-            Metric for calculating diversity of the subset "Gini", "Entropy" etc.
-            Default="Tanimoto".
-        random_seed : int, optional
-            Random seed for reproducibility. Default=42.
-        feature_type : str, optional
-            Type of features. Default=None.
-        mol_file : str, optional
-            Path to the file with molecules. Default=None.
-        feature_file : str, optional
-            Path to the file with features. Default=None.
-        num_selected : int, optional
-            Number of molecules to select. Default=None.
-        normalize_features : bool, optional
-            Normalize features or not. Default=False.
-        arr_dist : numpy.ndarray, optional
-            Array of distances between molecules. Default=None.
+        arr: np.ndarray
+            Array of features if fun_distance is provided.
+            Otherwise, treated as distance matrix.
+        func_distance: callable
+            function for calculating the pairwise distance between instances of the array.
+            Default is None.
+        num_selected: int
+            number of points that need to be selected
+        labels: np.ndarray
+            labels for performing algorithm withing clusters.
+            Default is None.
+
+        Returns
+        -------
+        selected: list
+            list of ids of selected molecules
+        """
+        self.n_mols = arr.shape[0]
+        if func_distance is not None:
+            self.arr_dist = func_distance(arr)
+        else:
+            self.arr_dist = arr
+
+        if labels is not None:
+            unique_labels = np.unique(labels)
+            num_clusters = len(unique_labels)
+            selected_all = []
+            totally_used = []
+
+            amount_molecules = np.array(
+                [len(np.where(labels == unique_label)[0]) for unique_label in unique_labels])
+
+            n = (num_selected - len(selected_all)) // (num_clusters - len(totally_used))
+
+            while np.any(amount_molecules <= n):
+                for unique_label in unique_labels:
+                    if unique_label not in totally_used:
+                        cluster_ids = np.where(labels == unique_label)[0]
+                        if len(cluster_ids) <= n:
+                            selected_all.append(cluster_ids)
+                            totally_used.append(unique_label)
+
+                n = (num_selected - len(selected_all)) // (num_clusters - len(totally_used))
+                amount_molecules = np.delete(amount_molecules, totally_used)
+
+                warnings.warn(f"Number of molecules in one cluster is less than"
+                              f" {num_selected}/{num_clusters}.\nNumber of selected "
+                              f"molecules might be less than desired.\nIn order to avoid this "
+                              f"problem. Try to use less number of clusters")
+
+            for unique_label in unique_labels:
+                if unique_label not in totally_used:
+                    cluster_ids = np.where(labels == unique_label)[0]
+                    selected = self.select_from_cluster(arr, n, cluster_ids)
+                    selected_all.append(cluster_ids[selected])
+            return np.hstack(selected_all).flatten().tolist()
+        else:
+            selected = self.select_from_cluster(self.arr_dist, num_selected)
+            return selected
+
+    @staticmethod
+    @abstractmethod
+    def select_from_cluster(arr_dist, num_selected):
+        """
+        Algorithm for selecting points from cluster.
+
+        Parameters
+        ----------
+        arr_dist: np.ndarray
+            distance matrix for points that needs to be selected
+        num_selected: int
+            number of molecules that need to be selected
+
+        Returns
+        -------
+        selected: list
+            list of ids of selected molecules
 
         """
-        self.num_selected = num_selected
-        self.normalize_features = normalize_features
-        self.random_seed = random_seed
-
-        # feature loader if string is
-        # accepts string or pure path object
-        if features is not None:
-            if isinstance(features, (str, PurePath)):
-                self.features = feature_reader(file_name=features,
-                                               sep=sep,
-                                               engine=engine,
-                                               **kwargs)
-            else:
-                self.features = features
-            # normalize features
-            if normalize_features:
-                self.features = StandardScaler().fit_transform(self.features)
-        # feature is None
-        else:
-            self.features = None
-
-        # todo: current version only works for molecular descriptors
-        # pair-wise distance matrix
-        if arr_dist is None:
-            self.arr_dist = compute_distance_matrix(features=self.features,
-                                                    metric="euclidean")
-        else:
-            self.arr_dist = arr_dist
-
-        if self.features is None and self.arr_dist is None:
-            raise ValueError("Features or distance matrix must be provided")
-
-    # abstract method, because we want in to be in both child classes
-    @abstractmethod
-    def select(self):
-        """Select the subset molecules with optimal diversity."""
         pass
 
     # concrete method, because we want in to be in both child classes, and it should act
